requirements.txt
pandas==2.2.2
Pillow==10.3.0
streamlit==1.34.0
ultralytics==8.2.8
Flask==3.0.0
